Add corpus test
